requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
pillow==10.0.1
rembg==2.0.50
numpy==1.26.0
python-dotenv==1.0.0
torch>=2.0.0
transformers>=4.30.0